Fixed win build (std::pair replaced with boost::tuple) Fixed compiler warnings
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@860 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ include/Transmission/Vector/VectorAnopheles.h
@@ -251,7 +251,7 @@ private:
      * @return non human host ento availability;
      *
      */
-  double getNonHumanEntoAvailability(int populationSize, double relativeEntoAvailability);
+  double getNonHumanEntoAvailability(double populationSize, double relativeEntoAvailability);
 
   /** Proportion of host-seeking parous mosquitoes that have laid eggs same day*/
   double mosqLaidEggsSameDayProp;
